# Conflicts:
#	src/main/java/buravel/buravel/modules/account/AccountService.java

+++ src/main/java/buravel/buravel/modules/bookmark/BookmarkController.java
@@ -69,20 +69,23 @@ public class BookmarkController {
     }
 
     @DeleteMapping(value = "/{bookmark_id}")
-    public ResponseEntity deleteBookmark(@PathVariable Long bookmark_id, @CurrentUser Account account){
+    public ResponseEntity deleteBookmark(@PathVariable Long bookmark_id, @CurrentUser Account account) throws NotFoundException {
         URI location = linkTo(BookmarkController.class).withRel("getBookmarkList").toUri();
-        try{
-            bookmarkService.deleteBookmark(bookmark_id,account);
-            return ResponseEntity.ok(location);
-        } catch (NotFoundException e) {
-            return ResponseEntity.notFound().build(); //not found bookmark
-        } catch (RuntimeException e){
-            return ResponseEntity.badRequest().body(location); //invalid user
+        if (!bookmarkRepository.existsById(bookmark_id)) {
+            return ResponseEntity.notFound().build();
         }
+        bookmarkService.deleteBookmark(bookmark_id,account);
+        return ResponseEntity.ok(location);
+
     }
 
     @PatchMapping(value = "/{bookmark_id}")
-    public ResponseEntity modifyBookmarkTitle(@PathVariable Long bookmark_id,@RequestBody @Valid BookmarkDto bookmarkDto,@CurrentUser Account account,Errors errors){
+    public ResponseEntity modifyBookmarkTitle(@PathVariable Long bookmark_id,@RequestBody @Valid BookmarkDto bookmarkDto,@CurrentUser Account account,Errors errors) throws NotFoundException{
+
+        if (!bookmarkRepository.existsById(bookmark_id)) {
+            return ResponseEntity.notFound().build();
+        }
+
         if(errors.hasErrors()) {
             EntityModel<Errors> error1 = ErrorResource.modelOf(errors);
             return ResponseEntity.badRequest().body(error1);
@@ -93,20 +96,15 @@ public class BookmarkController {
             EntityModel<Errors> error2 = ErrorResource.modelOf(errors);
             return ResponseEntity.badRequest().body(error2);
         }
+
         URI location = linkTo(BookmarkController.class).withRel("getBookmarkList").toUri();
-        try{
-            BookmarkResponseDto bookmarkResponseDto = bookmarkService.modifyBookmark(bookmark_id,bookmarkDto,account);
-            EntityModel<BookmarkResponseDto> responseResource = BookmarkResource.modelOf(bookmarkResponseDto);
+        BookmarkResponseDto bookmarkResponseDto = bookmarkService.modifyBookmark(bookmark_id,bookmarkDto,account);
+        EntityModel<BookmarkResponseDto> responseResource = BookmarkResource.modelOf(bookmarkResponseDto);
 
-            responseResource.add(linkTo(BookmarkController.class).withRel("getBookmarkList"));
-            responseResource.add(linkTo(BookmarkController.class).slash(bookmarkResponseDto.getId()).withRel("deleteBookmark"));
-            responseResource.add(linkTo(BookmarkController.class).slash(bookmarkResponseDto.getId()).withRel("modifyBookmark"));
+        responseResource.add(linkTo(BookmarkController.class).withRel("getBookmarkList"));
+        responseResource.add(linkTo(BookmarkController.class).slash(bookmarkResponseDto.getId()).withRel("deleteBookmark"));
+        responseResource.add(linkTo(BookmarkController.class).slash(bookmarkResponseDto.getId()).withRel("modifyBookmark"));
 
-            return ResponseEntity.ok(responseResource);
-        } catch (NotFoundException e) {
-            return ResponseEntity.notFound().build();
-        } catch (RuntimeException e){
-            return ResponseEntity.badRequest().body(location);
-        }
+        return ResponseEntity.ok(location);
     }
 }
